chore(workspace): clean clippy warnings and fix TUI snapshot failures
### Fixed
- Resolve clippy `overly_complex_bool_expr` deny error in `navi-openai/src/providers/openai.rs`.
- Fix flaky `tool_approval_pending_in_chat` screenshot by resetting request usage in the test harness for deterministic elapsed time.
- Update `modal_thinking_80x24` snapshot to include `xhigh` and `off` effort levels.

### Changed
- Box `ApiMeta` inside `ModelStreamEvent` to resolve `large_enum_variant` warning and reduce stream event size.
- Apply `cargo clippy --fix --workspace` for machine-applicable cleanups across the workspace.
- Reformat with `cargo fmt --all`.

+++ crates/navi-core/src/memory/embedding.rs
@@ -189,11 +189,7 @@ mod candle_embedder {
 
             // Create input_ids tensor [1, seq_len] as u32
             let input_ids_tensor = Tensor::from_slice(
-                input_ids
-                    .iter()
-                    .map(|&v| v as u32)
-                    .collect::<Vec<_>>()
-                    .as_slice(),
+                input_ids.iter().map(|&v| v).collect::<Vec<_>>().as_slice(),
                 (1, input_ids.len()),
                 &self.device,
             )?;
@@ -202,7 +198,7 @@ mod candle_embedder {
             let attention_mask_tensor = Tensor::from_slice(
                 attention_mask
                     .iter()
-                    .map(|&v| v as u32)
+                    .map(|&v| v)
                     .collect::<Vec<_>>()
                     .as_slice(),
                 (1, attention_mask.len()),
